+++ src/test/kotlin/stringCalculator/ValidatorTest.kt
@@ -1,0 +1,10 @@
+package stringCalculator
+
+import org.junit.jupiter.api.Test
+
+internal class ValidatorTest {
+
+    @Test
+    fun isValidInput() {
+    }
+}
